Avoid repeatedly allocating memory in Exchange operator
When a SerializedPage is sent to Exchange, it may contain multiple
IOBufs. The Exchange operator would grow the memory gradually, causing
multiple unnecessary memory allocations. This commit adds a new Http
header "X-Presto-Buffer-Total-Rows", so that the Exchange operator has
a way to know the total number of rows in the SerializedPage and
allocate memory with enough capacity in one allocation. The
ExchangeBenchmark shows up to 33% improvements in number of
allocations, up to 60% reduction on peak memory usage, and up to 90%
CPU time reduction on the isBlocked() calls for the Exchange operator.
The total wall time shows small improvements too.